issue #3: split up Accessory.js into SensorAccessory and ActorAccessory
